fix(data-warehouse): address CDC repair review findings
- dedupe external-data-sources-repair-cdc-create in tools.yaml: a rebase merged
  the scaffold's enabled:false stub alongside the configured entry, and the
  yaml parser resolves duplicate keys against the tool being enabled; keep one
  entry in its alphabetical slot and regenerate
- move cdc_qualified_table_name to cdc/naming.py so steady-state extraction no
  longer imports the repair module for a naming helper
- re-assert cdc enabled inside repair_cdc_source so future facade callers
  can't repair a CDC-disabled source
