+++ products/warehouse_sources/backend/temporal/data_imports/cdc/activities.py
@@ -56,7 +56,7 @@ from products.warehouse_sources.backend.temporal.data_imports.cdc.errors import 
     CDCSchemaMergeError,
     classify_cdc_error,
 )
-from products.warehouse_sources.backend.temporal.data_imports.cdc.repair import cdc_qualified_table_name
+from products.warehouse_sources.backend.temporal.data_imports.cdc.naming import cdc_qualified_table_name
 from products.warehouse_sources.backend.temporal.data_imports.cdc.types import ChangeEvent
 from products.warehouse_sources.backend.temporal.data_imports.pipelines.helpers import resolve_table_and_folder_names
 from products.warehouse_sources.backend.temporal.data_imports.pipelines.pipeline_v3.kafka.common import SyncTypeLiteral
